refactor: reorganize components and update imports
