+++ frontend/src/Components/NavBar.tsx
@@ -1,4 +1,4 @@
-import logo from './assets/logo.png';
+import logo from '../assets/logo.png';
 
 export default function NavBar() {
     return (
